requirements.txt
streamlit
python-dotenv
openai
streamlit-webrtc
websockets
pyaudio
pydub
reportlab